requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11
django-braces==1.11.0
django-cors-headers==2.0.2
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.6.2
oauthlib==2.0.1
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
